+++ niryo_robot_modbus/src/niryo_robot_modbus/mapping/coils.py
@@ -55,21 +55,20 @@ class ToolEquippedEntry(ABCRegisterEntry):
 @slave_context.coil
 class ToolActuationEntry(ABCRegisterEntry):
 
-    def __init__(self, ros_wrapper: NiryoRosWrapper):
-        super().__init__(ros_wrapper)
-        self._is_open = False
-        try:
-            specs = ros_wrapper.get_gripper_specs()
-        except NiryoRosWrapperException:
-            return
-        close_position = specs[0][0]
-        open_position = specs[0][1]
-        current_position = ros_wrapper.get_current_tool_position()
-        # Naive way to determine if the gripper is open or closed
-        self._is_open = abs(current_position - open_position) < abs(current_position - close_position)
+    def _grasp(self):
+        self._ros_wrapper.grasp_with_tool(speed=CommonStore.gripper_close_speed,
+                                          max_torque_percentage=CommonStore.gripper_close_max_torque,
+                                          hold_torque_percentage=CommonStore.gripper_close_hold_torque,
+                                          pin_id=PinID.DO4)
+
+    def _release(self):
+        self._ros_wrapper.release_with_tool(speed=CommonStore.gripper_open_speed,
+                                            max_torque_percentage=CommonStore.gripper_open_max_torque,
+                                            hold_torque_percentage=CommonStore.gripper_open_hold_torque,
+                                            pin_id=PinID.DO4)
 
     def get(self) -> bool:
-        return self._is_open
+        return self._ros_wrapper.is_tool_open()
 
     def set(self, value: bool) -> None:
         tool_id = self._ros_wrapper.get_current_tool_id()
@@ -77,25 +76,10 @@ class ToolActuationEntry(ABCRegisterEntry):
         if tool_id == ToolID.NONE:
             return
 
-        action_is_grasp = value
-        tool_is_gripper = tool_id in [ToolID.GRIPPER_1, ToolID.GRIPPER_2, ToolID.GRIPPER_3, ToolID.GRIPPER_4]
-
-        if action_is_grasp:
-            if tool_is_gripper:
-                self._ros_wrapper.close_gripper(CommonStore.gripper_close_speed,
-                                                CommonStore.gripper_close_max_torque,
-                                                CommonStore.gripper_close_hold_torque)
-            else:
-                self._ros_wrapper.grasp_with_tool(PinID.DO4)
-            self._is_open = False
+        if value:
+            self._grasp()
         else:
-            if tool_is_gripper:
-                self._ros_wrapper.open_gripper(CommonStore.gripper_open_speed,
-                                               CommonStore.gripper_open_max_torque,
-                                               CommonStore.gripper_open_hold_torque)
-            else:
-                self._ros_wrapper.release_with_tool(PinID.DO4)
-            self._is_open = True
+            self._release()
 
 
 @slave_context.coil